moving to cmake
