+++ MonSTRConfig.h.in
@@ -1,0 +1,2 @@
+#define MonSTR_VERSION_MAJOR @MonSTR_VERSION_MAJOR@
+#define MonSTR_VERSION_MINOR @MonSTR_VERSION_MINOR@
