+++ monstr/common.h
