+++ monstr/mathops.h
