+++ monstr/options.cpp
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with STRDenovoTools.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "src/options.h"
+#include "options.h"
 
 using namespace std;
