+++ monstr/vcf_reader.h
@@ -26,7 +26,7 @@ along with STRDenovoTools.  If not, see <http://www.gnu.org/licenses/>.
 #include <string>
 #include <vector>
 
-#include "src/common.h"
+#include "common.h"
 #include "htslib/bgzf.h"
 #include "htslib/tbx.h"
 #include "htslib/vcf.h"
