+++ pg_pedigree.fam
@@ -1,0 +1,3 @@
+1463 NA12878 NA12892 NA12891 2 -9
+1463 NA12891 0 0 1 -9
+1463 NA12892 0 0 2 -9
